chore: regroup flatten functions (#645)
