+++ src/services/filters-parser.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import Interface, { BaseFiltersParser, BaseOperatorDateParser } from 'forest-express';
 import { NoMatchingOperatorError, InvalidFiltersFormatError } from './errors';
 import utils from '../utils/schema';
-import FieldsFlattener from './fields-flattener';
+import Flattener from './flattener';
 
 const AGGREGATOR_OPERATORS = ['and', 'or'];
 
@@ -90,7 +90,7 @@ function FiltersParser(model, timezone, options) {
     const formatedField = this.formatField(condition.field);
 
     return {
-      [FieldsFlattener.unflattenFieldName(formatedField)]: await this.formatOperatorValue(
+      [Flattener.unflattenFieldName(formatedField)]: await this.formatOperatorValue(
         condition.field,
         condition.operator,
         condition.value,
@@ -227,7 +227,7 @@ function FiltersParser(model, timezone, options) {
     const newCondition = {
       operator: condition.operator,
       value: condition.value,
-      field: FieldsFlattener.unflattenFieldName(subFieldName),
+      field: Flattener.unflattenFieldName(subFieldName),
     };
     const query = await subModelFilterParser.perform(JSON.stringify(newCondition));
     const [, referencedKey] = field.reference.split('.');
@@ -235,7 +235,7 @@ function FiltersParser(model, timezone, options) {
     const subModelIds = subModelRecords.map((record) => record[referencedKey]);
 
     const resultCondition = {
-      field: FieldsFlattener.unflattenFieldName(fieldName),
+      field: Flattener.unflattenFieldName(fieldName),
       operator: 'in',
       value: subModelIds,
     };
@@ -243,7 +243,7 @@ function FiltersParser(model, timezone, options) {
       return {
         aggregator: 'or',
         conditions: [{
-          field: FieldsFlattener.unflattenFieldName(fieldName),
+          field: Flattener.unflattenFieldName(fieldName),
           operator: 'blank',
           value: null,
         }, resultCondition],
